Devtask/5 service employee

+++ internal/domain/staff.go
@@ -3,6 +3,6 @@ package domain
 type Employee struct {
 	ID      uint64
 	Name    string
-	PhotoID int
+	PhotoID uint64
 	Meta    map[string]interface{}
 }
